+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Image 0.31.0
+
+This is the changelog for Image version 0.31.0 released on May 7th, 2023.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-image/image/tags)
+
+### Bug Fixes
+
+* Fix resolution of the image returned from `Image.replace_color/2`
+
+### Enhancements
+
+* Add `:blend` option to `Image.if_then_else/4`. When set to true with smoothly blend between the `then` and `else`` images. The default is `false`.
+
 ## Image 0.30.0
 
 This is the changelog for Image version 0.30.0 released on April 20th, 2023.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-image/image/tags)